chore: 도메인 구조->계층형 구조로 패키지 구조 변경
