+++ src/main/java/gift/controller/ProductController.java
@@ -1,0 +1,82 @@
+package gift.controller;
+
+import gift.dto.ProductRequest;
+import gift.dto.ProductResponse;
+import gift.entity.Product;
+import gift.service.ProductService;
+import jakarta.validation.Valid;
+import java.util.List;
+import java.util.Optional;
+import java.util.stream.Collectors;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Slice;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+
+
+@RestController
+@RequestMapping("api/products")
+public class ProductController {
+
+    private final ProductService productService;
+
+    public ProductController(ProductService productService) {
+        this.productService = productService;
+    }
+
+    @GetMapping
+    public ResponseEntity<List<ProductResponse>> getAllProducts() {
+        List<Product> products = productService.findAll();
+        List<ProductResponse> response = products.stream()
+            .map(ProductResponse::from)
+            .collect(Collectors.toList());
+        return ResponseEntity.ok(response);
+    }
+
+    @GetMapping("/paged")
+    public ResponseEntity<Slice<ProductResponse>> getPagedProducts(
+        @RequestParam(defaultValue = "0") int page,
+        @RequestParam(defaultValue = "5") int size) {
+        Pageable pageable = PageRequest.of(page, size);
+        Slice<Product> productsPage = productService.findAll(pageable);
+        Slice<ProductResponse> responsePage = productsPage.map(ProductResponse::from);
+        return ResponseEntity.ok(responsePage);
+    }
+
+    @GetMapping("/{id}")
+    public ResponseEntity<ProductResponse> getProduct(@PathVariable Long id) {
+        Optional<Product> product = productService.findById(id);
+        if (product.isPresent()) {
+            ProductResponse productResponse = ProductResponse.from(product.get());
+            return ResponseEntity.ok(productResponse);
+        }
+        return ResponseEntity.status(204).build();
+    }
+
+    @PostMapping
+    public ResponseEntity<ProductResponse> addProduct(
+        @Valid @RequestBody ProductRequest productRequest) {
+        ProductResponse productResponse = productService.addProduct(productRequest);
+        return ResponseEntity.status(201).body(productResponse);
+    }
+
+    @PutMapping("/{id}")
+    public ResponseEntity<ProductResponse> updateProduct(@PathVariable Long id,
+        @Valid @RequestBody ProductRequest updatedProductRequest) {
+        if (!productService.findById(id).isPresent()) {
+            return ResponseEntity.status(204).build();
+        }
+        ProductResponse productResponse = productService.updateProduct(id, updatedProductRequest);
+        return ResponseEntity.ok(productResponse);
+    }
+
+    @DeleteMapping("/{id}")
+    public ResponseEntity<Void> deleteProduct(@PathVariable Long id) {
+        if (!productService.findById(id).isPresent()) {
+            return ResponseEntity.status(204).build();
+        }
+        productService.deleteById(id);
+        return ResponseEntity.ok().build();
+    }
+}
