+++ src/main/java/gift/dto/MemberRequest.java
@@ -1,0 +1,15 @@
+package gift.dto;
+
+public class MemberRequest {
+
+    private String email;
+    private String password;
+
+    public String getEmail() {
+        return email;
+    }
+
+    public String getPassword() {
+        return password;
+    }
+}
