+++ src/main/java/gift/dto/WishRequest.java
@@ -1,0 +1,17 @@
+package gift.dto;
+
+import gift.entity.Product;
+
+public class WishRequest {
+
+    private Product product;
+
+    public Product getProduct() {
+        return product;
+    }
+
+    public void setProduct(Product product) {
+        this.product = product;
+    }
+
+}
