--- src/main/java/gift/member/dto/MemberRequest.java
@@ -1,15 +1,0 @@
-package gift.member.dto;
-
-public class MemberRequest {
-
-    private String email;
-    private String password;
-
-    public String getEmail() {
-        return email;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-}
