--- src/main/java/gift/member/dto/MemberResponse.java
@@ -1,21 +1,0 @@
-package gift.member.dto;
-
-public class MemberResponse {
-
-    private Long id;
-    private String email;
-
-    public MemberResponse(Long id, String email) {
-        this.id = id;
-        this.email = email;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-}
