--- src/main/java/gift/member/repository/MemberRepository.java
@@ -1,12 +1,0 @@
-package gift.member.repository;
-
-import gift.member.model.Member;
-import java.util.Optional;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface MemberRepository extends JpaRepository<Member, Long> {
-
-    Optional<Member> findByEmail(String email);
-}
