--- src/main/java/gift/product/model/Product.java
@@ -1,68 +1,0 @@
-package gift.product.model;
-
-import gift.product.dto.ProductRequest;
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
-
-@Entity
-@Table(name = "product")
-public class Product {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    private String name;
-
-    @Column(nullable = false)
-    private int price;
-
-    @Column(nullable = false)
-    private String imgUrl;
-
-    public Product() {
-    }
-
-    public Product(String name, int price, String imgUrl) {
-        this.name = name;
-        this.price = price;
-        this.imgUrl = imgUrl;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public void updateId(Long id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public int getPrice() {
-        return price;
-    }
-
-    public void setPrice(int price) {
-        this.price = price;
-    }
-
-    public String getImgUrl() {
-        return imgUrl;
-    }
-
-    public void setImgUrl(String imgUrl) {
-        this.imgUrl = imgUrl;
-    }
-
-}
