--- src/main/java/gift/product/repository/ProductRepository.java
@@ -1,10 +1,0 @@
-package gift.product.repository;
-
-import gift.product.model.Product;
-import org.springframework.stereotype.Repository;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-@Repository
-public interface ProductRepository extends JpaRepository<Product, Long> {
-
-}
