+++ src/main/java/gift/repository/MemberRepository.java
@@ -1,0 +1,12 @@
+package gift.repository;
+
+import gift.entity.Member;
+import java.util.Optional;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface MemberRepository extends JpaRepository<Member, Long> {
+
+    Optional<Member> findByEmail(String email);
+}
