+++ src/main/java/gift/repository/ProductRepository.java
@@ -1,0 +1,10 @@
+package gift.repository;
+
+import gift.entity.Product;
+import org.springframework.stereotype.Repository;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+@Repository
+public interface ProductRepository extends JpaRepository<Product, Long> {
+
+}
