--- src/main/java/gift/wishlist/repository/WishRepository.java
@@ -1,18 +1,0 @@
-package gift.wishlist.repository;
-
-import gift.wishlist.model.Wish;
-import java.util.List;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface WishRepository extends JpaRepository<Wish, Long> {
-
-    List<Wish> findByMemberIdAndProductId(Long memberId, Long productId);
-
-    List<Wish> findByMemberId(Long memberId);
-
-    Page<Wish> findByMemberId(Long memberId, Pageable pageable);
-}
