+++ src/test/java/gift/member/repository/MemberRepositoryTest.java
@@ -2,7 +2,8 @@ package gift.member.repository;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import gift.member.model.Member;
+import gift.entity.Member;
+import gift.repository.MemberRepository;
 import java.util.Optional;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
