+++ src/test/java/gift/product/repository/ProductRepositoryTest.java
@@ -2,7 +2,8 @@ package gift.product.repository;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import gift.product.model.Product;
+import gift.entity.Product;
+import gift.repository.ProductRepository;
 import java.util.List;
 import java.util.Optional;
 import org.junit.jupiter.api.BeforeEach;
