+++ src/test/java/gift/wishlist/repository/WishRepositoryTest.java
@@ -2,11 +2,12 @@ package gift.wishlist.repository;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import gift.member.model.Member;
-import gift.member.repository.MemberRepository;
-import gift.product.model.Product;
-import gift.product.repository.ProductRepository;
-import gift.wishlist.model.Wish;
+import gift.entity.Member;
+import gift.repository.MemberRepository;
+import gift.entity.Product;
+import gift.repository.ProductRepository;
+import gift.entity.Wish;
+import gift.repository.WishRepository;
 import jakarta.persistence.EntityManager;
 import jakarta.transaction.Transactional;
 import java.util.List;
